+++ app/src/main/java/com/rviannaoliveira/vreddit/core/data/DataManagerFactory.kt
@@ -1,4 +1,4 @@
-package com.rviannaoliveira.vreddit.data
+package com.rviannaoliveira.vreddit.core.data
 
 /**
  * Criado por rodrigo on 18/10/17.
