+++ app/src/main/java/com/rviannaoliveira/vreddit/core/data/repository/RedditRepositoryDataSource.kt
@@ -1,4 +1,4 @@
-package com.rviannaoliveira.vreddit.data.repository
+package com.rviannaoliveira.vreddit.core.data.repository
 
 import com.rviannaoliveira.vreddit.modal.CommentData
 import com.rviannaoliveira.vreddit.modal.NewsData
