+++ app/src/main/java/com/rviannaoliveira/vreddit/core/extensions/ActivityExtensions.kt
