+++ app/src/main/java/com/rviannaoliveira/vreddit/core/global/ConstantsParceable.kt
@@ -1,4 +1,4 @@
-package com.rviannaoliveira.vreddit.global
+package com.rviannaoliveira.vreddit.core.global
 
 /**
  * Criado por rodrigo on 21/10/17.
