+++ app/src/main/java/com/rviannaoliveira/vreddit/core/global/ProgressViewInterface.kt
@@ -1,4 +1,4 @@
-package com.rviannaoliveira.vreddit.global
+package com.rviannaoliveira.vreddit.core.global
 
 /**
  * Criado por rodrigo on 18/10/17.
