+++ app/src/main/java/com/rviannaoliveira/vreddit/core/model/CommentData.kt
