+++ app/src/main/java/com/rviannaoliveira/vreddit/core/model/NewsChildrenDataNvl2Response.kt
