+++ app/src/main/java/com/rviannaoliveira/vreddit/core/util/DefaultData.kt
