+++ app/src/main/java/com/rviannaoliveira/vreddit/detail/DetailActivity.kt
@@ -4,9 +4,9 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
 import com.rviannaoliveira.vreddit.R
+import com.rviannaoliveira.vreddit.core.global.ConstantsParceable
 import com.rviannaoliveira.vreddit.extensions.closeActivityWithSlideTransictionCloseDownActivity
 import com.rviannaoliveira.vreddit.extensions.showFragment
-import com.rviannaoliveira.vreddit.global.ConstantsParceable
 import com.rviannaoliveira.vreddit.modal.NewsData
 
 /**
