+++ app/src/test/java/com/rviannaoliveira/vreddit/presenter/DetailPresenterTest.kt
@@ -3,7 +3,7 @@ package com.rviannaoliveira.vreddit.presenter
 import com.nhaarman.mockito_kotlin.verify
 import com.nhaarman.mockito_kotlin.whenever
 import com.rviannaoliveira.vreddit.BuildConfig
-import com.rviannaoliveira.vreddit.data.DataManagerInterface
+import com.rviannaoliveira.vreddit.core.data.DataManagerInterface
 import com.rviannaoliveira.vreddit.detail.DetailInterface
 import com.rviannaoliveira.vreddit.detail.DetailPresenterImpl
 import com.rviannaoliveira.vreddit.modal.CommentData
